+++ routes/web.php
@@ -20,3 +20,4 @@ Route::get('dbtest', function () {
     return $data;
 });
 
+